requirements.txt
pandas
numpy
scikit-learn
great_expectations
google-generativeai
python-dotenv
requests
joblib
streamlit
